yt_shorts_bot/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
yt-dlp>=2024.1.0
imageio-ffmpeg>=0.4.9
ultralytics>=8.0.0

